refactor: seller 도메인 로직 분리
