+++ src/features/seller/resolvers/seller-order-query.resolver.ts
@@ -1,0 +1,37 @@
+import { UseGuards } from '@nestjs/common';
+import { Args, Query, Resolver } from '@nestjs/graphql';
+
+import { CurrentUser, JwtAuthGuard, type JwtUser } from '../../../global/auth';
+import { SellerService } from '../seller.service';
+import type { SellerOrderListInput } from '../types/seller-input.type';
+import type {
+  SellerCursorConnection,
+  SellerOrderDetailOutput,
+  SellerOrderSummaryOutput,
+} from '../types/seller-output.type';
+
+import { parseAccountId, parseId } from './seller-resolver.utils';
+
+@Resolver('Query')
+@UseGuards(JwtAuthGuard)
+export class SellerOrderQueryResolver {
+  constructor(private readonly sellerService: SellerService) {}
+
+  @Query('sellerOrderList')
+  sellerOrderList(
+    @CurrentUser() user: JwtUser,
+    @Args('input', { nullable: true }) input?: SellerOrderListInput,
+  ): Promise<SellerCursorConnection<SellerOrderSummaryOutput>> {
+    const accountId = parseAccountId(user);
+    return this.sellerService.sellerOrderList(accountId, input);
+  }
+
+  @Query('sellerOrder')
+  sellerOrder(
+    @CurrentUser() user: JwtUser,
+    @Args('orderId') orderId: string,
+  ): Promise<SellerOrderDetailOutput> {
+    const accountId = parseAccountId(user);
+    return this.sellerService.sellerOrder(accountId, parseId(orderId));
+  }
+}
