+++ src/features/seller/resolvers/seller-product-query.resolver.ts
@@ -1,0 +1,36 @@
+import { UseGuards } from '@nestjs/common';
+import { Args, Query, Resolver } from '@nestjs/graphql';
+
+import { CurrentUser, JwtAuthGuard, type JwtUser } from '../../../global/auth';
+import { SellerService } from '../seller.service';
+import type { SellerProductListInput } from '../types/seller-input.type';
+import type {
+  SellerCursorConnection,
+  SellerProductOutput,
+} from '../types/seller-output.type';
+
+import { parseAccountId, parseId } from './seller-resolver.utils';
+
+@Resolver('Query')
+@UseGuards(JwtAuthGuard)
+export class SellerProductQueryResolver {
+  constructor(private readonly sellerService: SellerService) {}
+
+  @Query('sellerProducts')
+  sellerProducts(
+    @CurrentUser() user: JwtUser,
+    @Args('input', { nullable: true }) input?: SellerProductListInput,
+  ): Promise<SellerCursorConnection<SellerProductOutput>> {
+    const accountId = parseAccountId(user);
+    return this.sellerService.sellerProducts(accountId, input);
+  }
+
+  @Query('sellerProduct')
+  sellerProduct(
+    @CurrentUser() user: JwtUser,
+    @Args('productId') productId: string,
+  ): Promise<SellerProductOutput> {
+    const accountId = parseAccountId(user);
+    return this.sellerService.sellerProduct(accountId, parseId(productId));
+  }
+}
