+++ src/features/seller/resolvers/seller-store-mutation.resolver.ts
@@ -1,0 +1,95 @@
+import { UseGuards } from '@nestjs/common';
+import { Args, Mutation, Resolver } from '@nestjs/graphql';
+
+import { CurrentUser, JwtAuthGuard, type JwtUser } from '../../../global/auth';
+import { SellerService } from '../seller.service';
+import type {
+  SellerUpdatePickupPolicyInput,
+  SellerUpdateStoreBasicInfoInput,
+  SellerUpsertStoreBusinessHourInput,
+  SellerUpsertStoreDailyCapacityInput,
+  SellerUpsertStoreSpecialClosureInput,
+} from '../types/seller-input.type';
+import type {
+  SellerStoreBusinessHourOutput,
+  SellerStoreDailyCapacityOutput,
+  SellerStoreOutput,
+  SellerStoreSpecialClosureOutput,
+} from '../types/seller-output.type';
+
+import { parseAccountId, parseId } from './seller-resolver.utils';
+
+@Resolver('Mutation')
+@UseGuards(JwtAuthGuard)
+export class SellerStoreMutationResolver {
+  constructor(private readonly sellerService: SellerService) {}
+
+  @Mutation('sellerUpdateStoreBasicInfo')
+  sellerUpdateStoreBasicInfo(
+    @CurrentUser() user: JwtUser,
+    @Args('input') input: SellerUpdateStoreBasicInfoInput,
+  ): Promise<SellerStoreOutput> {
+    const accountId = parseAccountId(user);
+    return this.sellerService.sellerUpdateStoreBasicInfo(accountId, input);
+  }
+
+  @Mutation('sellerUpsertStoreBusinessHour')
+  sellerUpsertStoreBusinessHour(
+    @CurrentUser() user: JwtUser,
+    @Args('input') input: SellerUpsertStoreBusinessHourInput,
+  ): Promise<SellerStoreBusinessHourOutput> {
+    const accountId = parseAccountId(user);
+    return this.sellerService.sellerUpsertStoreBusinessHour(accountId, input);
+  }
+
+  @Mutation('sellerUpsertStoreSpecialClosure')
+  sellerUpsertStoreSpecialClosure(
+    @CurrentUser() user: JwtUser,
+    @Args('input') input: SellerUpsertStoreSpecialClosureInput,
+  ): Promise<SellerStoreSpecialClosureOutput> {
+    const accountId = parseAccountId(user);
+    return this.sellerService.sellerUpsertStoreSpecialClosure(accountId, input);
+  }
+
+  @Mutation('sellerDeleteStoreSpecialClosure')
+  sellerDeleteStoreSpecialClosure(
+    @CurrentUser() user: JwtUser,
+    @Args('closureId') closureId: string,
+  ): Promise<boolean> {
+    const accountId = parseAccountId(user);
+    return this.sellerService.sellerDeleteStoreSpecialClosure(
+      accountId,
+      parseId(closureId),
+    );
+  }
+
+  @Mutation('sellerUpdatePickupPolicy')
+  sellerUpdatePickupPolicy(
+    @CurrentUser() user: JwtUser,
+    @Args('input') input: SellerUpdatePickupPolicyInput,
+  ): Promise<SellerStoreOutput> {
+    const accountId = parseAccountId(user);
+    return this.sellerService.sellerUpdatePickupPolicy(accountId, input);
+  }
+
+  @Mutation('sellerUpsertStoreDailyCapacity')
+  sellerUpsertStoreDailyCapacity(
+    @CurrentUser() user: JwtUser,
+    @Args('input') input: SellerUpsertStoreDailyCapacityInput,
+  ): Promise<SellerStoreDailyCapacityOutput> {
+    const accountId = parseAccountId(user);
+    return this.sellerService.sellerUpsertStoreDailyCapacity(accountId, input);
+  }
+
+  @Mutation('sellerDeleteStoreDailyCapacity')
+  sellerDeleteStoreDailyCapacity(
+    @CurrentUser() user: JwtUser,
+    @Args('capacityId') capacityId: string,
+  ): Promise<boolean> {
+    const accountId = parseAccountId(user);
+    return this.sellerService.sellerDeleteStoreDailyCapacity(
+      accountId,
+      parseId(capacityId),
+    );
+  }
+}
