+++ src/features/seller/seller.service.spec.ts
@@ -1,31 +1,48 @@
-import {
-  BadRequestException,
-  ForbiddenException,
-  NotFoundException,
-} from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
-import { OrderStatus } from '@prisma/client';
 
-import { SellerRepository } from './repositories/seller.repository';
 import { SellerService } from './seller.service';
+import { SellerContentService } from './services/seller-content.service';
+import { SellerConversationService } from './services/seller-conversation.service';
+import { SellerOrderService } from './services/seller-order.service';
+import { SellerProductService } from './services/seller-product.service';
+import { SellerStoreService } from './services/seller-store.service';
 
 describe('SellerService', () => {
   let service: SellerService;
-  let repo: jest.Mocked<SellerRepository>;
+  let storeService: jest.Mocked<SellerStoreService>;
+  let orderService: jest.Mocked<SellerOrderService>;
 
   beforeEach(async () => {
-    repo = {
-      findSellerAccountContext: jest.fn(),
-      findOrderDetailByStore: jest.fn(),
-      updateOrderStatusBySeller: jest.fn(),
-    } as unknown as jest.Mocked<SellerRepository>;
+    storeService = {
+      sellerMyStore: jest.fn(),
+    } as unknown as jest.Mocked<SellerStoreService>;
+
+    orderService = {
+      sellerUpdateOrderStatus: jest.fn(),
+    } as unknown as jest.Mocked<SellerOrderService>;
 
     const module: TestingModule = await Test.createTestingModule({
       providers: [
         SellerService,
         {
-          provide: SellerRepository,
-          useValue: repo,
+          provide: SellerStoreService,
+          useValue: storeService,
+        },
+        {
+          provide: SellerProductService,
+          useValue: {},
+        },
+        {
+          provide: SellerOrderService,
+          useValue: orderService,
+        },
+        {
+          provide: SellerConversationService,
+          useValue: {},
+        },
+        {
+          provide: SellerContentService,
+          useValue: {},
         },
       ],
     }).compile();
@@ -33,70 +50,26 @@ describe('SellerService', () => {
     service = module.get<SellerService>(SellerService);
   });
 
-  it('판매자 계정이 아니면 ForbiddenException을 던져야 한다', async () => {
-    repo.findSellerAccountContext.mockResolvedValue({
-      id: BigInt(1),
-      account_type: 'USER',
-      status: 'ACTIVE',
-      store: { id: BigInt(100) },
-    } as never);
+  it('sellerMyStore는 storeService로 위임해야 한다', async () => {
+    await service.sellerMyStore(BigInt(11));
 
-    await expect(service.sellerMyStore(BigInt(1))).rejects.toThrow(
-      ForbiddenException,
-    );
+    expect(storeService.sellerMyStore).toHaveBeenCalledWith(BigInt(11));
   });
 
-  it('주문 상태 전이가 잘못되면 BadRequestException을 던져야 한다', async () => {
-    repo.findSellerAccountContext.mockResolvedValue({
-      id: BigInt(1),
-      account_type: 'SELLER',
-      status: 'ACTIVE',
-      store: { id: BigInt(100) },
-    } as never);
+  it('sellerUpdateOrderStatus는 orderService로 위임해야 한다', async () => {
+    await service.sellerUpdateOrderStatus(BigInt(11), {
+      orderId: '1',
+      toStatus: 'CONFIRMED',
+      note: null,
+    });
 
-    repo.findOrderDetailByStore.mockResolvedValue({
-      id: BigInt(10),
-      status: OrderStatus.SUBMITTED,
-      order_number: 'O-1',
-      account_id: BigInt(20),
-      pickup_at: new Date(),
-      buyer_name: '홍길동',
-      buyer_phone: '010-0000-0000',
-      subtotal_price: 1000,
-      discount_price: 0,
-      total_price: 1000,
-      submitted_at: new Date(),
-      confirmed_at: null,
-      made_at: null,
-      picked_up_at: null,
-      canceled_at: null,
-      created_at: new Date(),
-      updated_at: new Date(),
-      status_histories: [],
-      items: [],
-    } as never);
-
-    await expect(
-      service.sellerUpdateOrderStatus(BigInt(1), {
-        orderId: '10',
-        toStatus: 'MADE',
+    expect(orderService.sellerUpdateOrderStatus).toHaveBeenCalledWith(
+      BigInt(11),
+      {
+        orderId: '1',
+        toStatus: 'CONFIRMED',
         note: null,
-      }),
-    ).rejects.toThrow(BadRequestException);
-  });
-
-  it('주문이 없으면 NotFoundException을 던져야 한다', async () => {
-    repo.findSellerAccountContext.mockResolvedValue({
-      id: BigInt(1),
-      account_type: 'SELLER',
-      status: 'ACTIVE',
-      store: { id: BigInt(100) },
-    } as never);
-
-    repo.findOrderDetailByStore.mockResolvedValue(null);
-
-    await expect(service.sellerOrder(BigInt(1), BigInt(9999))).rejects.toThrow(
-      NotFoundException,
+      },
     );
   });
 });
